requirements.txt
flask
gunicorn
python-binance
psycopg2
Flask-SQLAlchemy
python-dotenv